review entity 수정
rating을 1~5사이 정수형이 될 수 있게 변경

ref #57

+++ src/main/java/com/rodemtree/yeyakitda/dto/ReviewDto.java
@@ -11,6 +11,6 @@ public record ReviewDto(
         String userNickname,
         List<String> imageUrls,
         String comment,
-        Float rating
+        Integer rating
 ) {
 }
